Moved cvss related functions to the calculator
